Refactor: rearrange composable structure
